+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(home)/_components/ContextMenu/ContextMenu.tsx
@@ -1,58 +1,23 @@
+'use client';
+
 import Typography from '@/components/Typography';
-import { useContextMenu } from '../../_provider/ContextMenuProvider';
 import CopyIcon from '@/icons/Copy.svg';
 import HashIcon from '@/icons/Hash.svg';
 import TrashIcon from '@/icons/Trash.svg';
 import { CHAT_TYPE } from '@/constants/chat';
 import clsx from 'clsx';
-import { useSnackBar } from '@/providers/SnackBarContext';
-import { useParams } from 'next/navigation';
-import { useMutationChatMessage, useMutationDeleteChatMessage } from '../../../../_hook/useChatMutation';
+import { ComponentPropsWithoutRef } from 'react';
+import { useContextMenuActions } from '../../_hooks/useContextMenuActions';
 
 const ContextMenu = () => {
-  const { id } = useParams();
-
-  const { contextMenuState, closeContextMenu } = useContextMenu();
-  const { openSnackBar } = useSnackBar();
-
-  const { mutate: mutateChatMessage } = useMutationChatMessage({
-    channel_id: Number(id)
-  });
-
-  const { mutateAsync: mutateDeleteChatMessage } = useMutationDeleteChatMessage({
-    channel_id: Number(id)
-  });
-
-  const deleteChat = async () => {
-    if (!contextMenuState.id) return;
-
-    await mutateDeleteChatMessage(contextMenuState.id);
-
-    openSnackBar({ message: '삭제가 완료되었어요' });
-
-    closeContextMenu();
-  };
-
-  const copyText = () => {
-    window.navigator.clipboard.writeText(contextMenuState.text || '').then(() => {
-      openSnackBar({ message: '복사가 완료되었어요' });
-    });
-
-    closeContextMenu();
-  };
-
-  const handleNotice = () => {
-    mutateChatMessage({ content: contextMenuState.text || '', type: CHAT_TYPE.notice });
-    closeContextMenu();
-  };
+  const { contextMenuState, closeContextMenu, copyText, deleteChat, handleNotice } = useContextMenuActions();
 
   if (!contextMenuState.isOpen) return null;
-
   const position = contextMenuState.isMe ? contextMenuState.position : contextMenuState.position + 40;
 
   return (
     <>
-      <ul
+      <div
         style={{ bottom: position }}
         className={clsx(
           'fixed rounded-[6px] bg-bgBackground1 shadow-[0px_1px_8px_0px_rgba(0,0,0,0.15)] z-50 w-[154px] flex p-4 gap-[20px] flex-col',
@@ -61,30 +26,26 @@ const ContextMenu = () => {
         )}
       >
         {contextMenuState.type === CHAT_TYPE.text && (
-          <li className="flex items-center justify-between" onClick={copyText}>
-            <Typography variant="Subtitle16px" color="grey900">
-              복사
-            </Typography>
-            <CopyIcon />
-          </li>
+          <ContextMenuButton title="복사" onClick={copyText} icon={<CopyIcon />} />
         )}
-        <li className="flex items-center justify-between" onClick={handleNotice}>
-          <Typography variant="Subtitle16px" color="grey900">
-            공지
-          </Typography>
-          <HashIcon />
-        </li>
-        {contextMenuState.isMe && (
-          <li className="flex items-center justify-between" onClick={deleteChat}>
-            <Typography variant="Subtitle16px" color="grey900">
-              삭제
-            </Typography>
-            <TrashIcon />
-          </li>
-        )}
-      </ul>
+        <ContextMenuButton title="공지" onClick={handleNotice} icon={<HashIcon />} />
+        {contextMenuState.isMe && <ContextMenuButton title="삭제" onClick={deleteChat} icon={<TrashIcon />} />}
+      </div>
       <div className="fixed top-0 left-0 w-full h-full z-40" onClick={closeContextMenu} />
     </>
+  );
+};
+
+type ContextMenuButtonProps = { title: string; icon: React.ReactNode } & ComponentPropsWithoutRef<'button'>;
+
+const ContextMenuButton = ({ title, icon, ...props }: ContextMenuButtonProps) => {
+  return (
+    <button type="button" className="flex items-center justify-between" {...props}>
+      <Typography variant="Subtitle16px" color="grey900">
+        {title}
+      </Typography>
+      {icon}
+    </button>
   );
 };
